pyproject.toml
# =============================================================================
# Project Configuration
# =============================================================================
# Literature Review Assistant
# =============================================================================

[project]
name = "litrev-assistant"
version = "1.0.0"
description = "AI-powered literature review assistant using multi-agent collaboration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Nilesh Mishra"}
]
keywords = [
    "ai",
    "literature-review",
    "research",
    "arxiv",
    "autogen",
    "multi-agent",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "autogen-agentchat>=0.4.0",
    "autogen-core>=0.4.0",
    "autogen-ext[openai]>=0.4.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "arxiv>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "streamlit>=1.28.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "types-requests>=2.31.0",
    "ipython>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/nileshmishra/litrev-assistant"
Documentation = "https://github.com/nileshmishra/litrev-assistant#readme"
Repository = "https://github.com/nileshmishra/litrev-assistant"

# =============================================================================
# Build System
# =============================================================================

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "frontend*"]

# =============================================================================
# Ruff Configuration
# =============================================================================

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.isort]
known-first-party = ["src", "frontend"]

# =============================================================================
# Mypy Configuration
# =============================================================================

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

# =============================================================================
# Coverage Configuration
# =============================================================================

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
